fix(recovery): verify digest and exact byte length
